% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexsimple.R
\name{nexsimple}
\alias{nexsimple}
\title{nexsimple: a specific workflow for getting a nexus phylogeny to look how i want it}
\usage{
nexsimple(nex, df, var)
}
\arguments{
\item{nex}{a nexus data set}

\item{df}{a data frame with a column that contains "Genus_species" for multiple species}

\item{var}{the variable of interest, usually the predictor, a column within df in quotes}
}
\description{
this function starts with a nexus file and a data frame of variables organized by species, and prunes both to match for easier phylogenetic analysis. It also returns some nice visuals.
}
\examples{
\dontrun{
nex <- read.nexus("gennex_output.nex")
df  <- read.csv("VarBySpecies.csv")
nexsimple(nex = nex, df = df, var = "CPP")
}
}
\keyword{nexus,}
\keyword{phylogeny,}
\keyword{visual}
