% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyreg.R
\name{phyreg}
\alias{phyreg}
\title{phyreg: run a single regression on a nexus form}
\usage{
phyreg(tree = tree, df = df, pvar = pvar, ovar = ovar)
}
\arguments{
\item{tree}{a nexus data set}

\item{df}{dataset with column 1}

\item{pvar}{predictor variable(s), in the standard format for linear models if multiple}

\item{ovar}{outcome variable(s)}
}
\description{
phyreg: run a single regression on a nexus form
}
\examples{
\dontrun{
tree <- read.nexus("gennex_output.nex") #needs to be in class phylo
df <- read.csv("VarBySpecies.csv")
phyreg(tree = tree, df = df, pvar = "CPP", ovar = "death.prop")
}
}
