% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gennex.R
\name{gennex}
\alias{gennex}
\title{gennex: from genetic data to nexus phylogeny}
\usage{
gennex(data = data, bs = bs, out = out)
}
\arguments{
\item{data}{a list of accession numbers for mitogenomes in GenBank (https://blast.ncbi.nlm.nih.gov/Blast.cgi?PROGRAM=blastn&PAGE_TYPE=BlastSearch&LINK_LOC=blasthome)}

\item{bs}{how many bootstrap samples do you want it to run}

\item{out}{one of the species names in the "data" list to be used as the outgroup}
}
\description{
This function pulls mitochondrial DNA from Gen Bank and creates a phylogenetic map
}
\examples{
gennex(data = c(Canis_lupus_familiaris = "NC_002008", Lemur_catta = "NC_004025",Galago_moholi = "KC757396", Saimiri_sciureus = "FJ785425", Callithrix_jacchus = "NC_025586"), bs = 100, out = "Canis_lupus_familiaris")
}
\keyword{genbank}
\keyword{mitodna,}
\keyword{phylogeny,}
